chore: split wifi_basic.dart into multiple files
