+++ packages/wifi_basic/lib/wifi_generations.dart
@@ -1,0 +1,3 @@
+part of 'wifi_basic.dart';
+
+enum WiFiGenerations { unknown, legacy, wifi4, wifi5, wifi6 }
