+++ packages/wifi_basic/lib/wifi_network_security.dart
@@ -1,0 +1,3 @@
+part of 'wifi_basic.dart';
+
+enum WiFiNetworkSecurity { unknown, none, wep, wpa, wpa2, wpa3 }
